I'm tired of wrestling with this.
We conditionally define random_bytes() and fallback to an
implementation that ALWAYS throws an Exception. However, in
libraries that consume random_compat, we want the type hint to
always be a string.

So, I'm done. Suppress the damn error (but still print it) and
let's move on with our lives.